- added testing endpoint to send predefined search eval job definition to supervisor for processing test - changed indices for judgement file parsing (make configurable!) - added route to ClusterNode startup - minor reformat - JobManagerActor/WorkManager: make explicit that WOrkManagerActor is job specific, reduce nr of WorkManagers per jobId to 1 per node - added JobMessages to pass general serializable definition instead of full ActorRunnable to reduce the pain of making everything serializable + providing implicit toRunnable conversion that can be executed on the respective node (temporary testing state, right now the search evaluation implicit returns a predefined runnable) - strategy renaming for JudgementHandlingStrategy - making thoughput actor optional in RequestProcessingFlows - SearchJobDefinitions for some test compositions for search eval jobs (temporary, its just a messy test hack now) - added test_judgements.txt - TestMessages: pass jobId - WorkManagerActor: added job messages for test pi calc job and search opt , making more explicit that one WorkManager is only for specific jobId

+++ src/main/scala/de/awagen/kolibri/base/domain/jobdefinitions/TestJobDefinitions.scala
@@ -87,7 +87,7 @@ object TestJobDefinitions {
     }
     JobMsgFactory.createActorRunnableJobCmd[Int, Int, Double, Double, Map[Tag, AggregateValue[Double]]](
       jobId = jobName,
-      nrThrows.toInt,
+      nrThrows,
       dataBatchGenerator = batchGenerator,
       transformerFlow = Flow.fromFunction[Int, ProcessingMessage[Double]](flowFunct),
       processingActorProps = None,
